+++ checkpoint/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validate merged and saved checkpoint params by comparing it against its
  `.zarray` counterpart.
 - Added private utilities to work with array fragments (`_src/arrays`).
+- Log `orbax-checkpoint version` when Checkpointer instances are created.
 
 ## [0.6.1] - 2024-08-22